models refactor
