--- src/main/java/model/Admin.java
@@ -1,7 +1,0 @@
-package model;
-
-public class Admin extends User {
-    public Admin(int id, String login, String password, String firstName, String lastName) {
-        super(id, login, password, firstName, lastName);
-    }
-}
