--- src/main/java/model/Artifact.java
@@ -1,7 +1,0 @@
-package model;
-
-public class Artifact extends Element {
-    public Artifact(int id, String name, String description, int value, int categoryId) {
-        super(id, name, description, value, categoryId);
-    }
-}
