--- src/main/java/model/Category.java
@@ -1,7 +1,0 @@
-package model;
-
-public class Category extends Group{
-    public Category(int id, String name) {
-        super(id, name);
-    }
-}
