--- src/main/java/model/Classroom.java
@@ -1,7 +1,0 @@
-package model;
-
-public class Classroom extends Group{
-    public Classroom(int id, String name) {
-        super(id, name);
-    }
-}
