--- src/main/java/model/Mentor.java
@@ -1,8 +1,0 @@
-package model;
-
-public class Mentor extends User {
-
-    public Mentor(int id, String login, String password, String firstName, String lastName, Integer studentId) {
-        super(id, login, password, firstName, lastName);
-    }
-}
