--- src/main/java/model/Quest.java
@@ -1,8 +1,0 @@
-package model;
-
-public class Quest extends Element {
-
-    public Quest(int id, String name, String description, int value, int categoryId) {
-        super(id, name, description, value, categoryId);
-    }
-}
