--- src/main/java/model/Team.java
@@ -1,7 +1,0 @@
-package model;
-
-public class Team extends Group {
-    public Team(int id, String name) {
-        super(id, name);
-    }
-}
