+++ src/main/java/model/elements/Artifact.java
@@ -1,0 +1,5 @@
+package model.elements;
+
+public class Artifact extends Element {
+
+}
