+++ src/main/java/model/elements/Quest.java
@@ -1,0 +1,5 @@
+package model.elements;
+
+public class Quest extends Element {
+
+}
