+++ src/main/java/model/groups/Category.java
@@ -1,0 +1,5 @@
+package model.groups;
+
+public class Category extends Group{
+
+}
