+++ src/main/java/model/groups/Classroom.java
@@ -1,0 +1,5 @@
+package model.groups;
+
+public class Classroom extends Group{
+
+}
