+++ src/main/java/model/groups/Group.java
@@ -1,13 +1,8 @@
-package model;
+package model.groups;
 
 public abstract class Group {
     private int id;
     private String name;
-
-    public Group(int id, String name){
-        this.id = id;
-        this.name = name;
-    }
 
     public int getId() {
         return id;
