+++ src/main/java/model/groups/Team.java
@@ -1,0 +1,5 @@
+package model.groups;
+
+public class Team extends Group {
+
+}
