+++ src/main/java/model/users/Admin.java
@@ -1,0 +1,5 @@
+package model.users;
+
+public class Admin extends User {
+
+}
