+++ src/main/java/model/users/Mentor.java
@@ -1,0 +1,5 @@
+package model.users;
+
+public class Mentor extends User {
+
+}
