+++ src/main/java/model/users/Student.java
@@ -1,0 +1,13 @@
+package model.users;
+
+public class Student extends User {
+    private int wallet;
+
+    public int getWallet() {
+        return wallet;
+    }
+
+    public void setWallet(int wallet) {
+        this.wallet = wallet;
+    }
+}
